Move base64 codecs file
And remove all references to architecture specific folders.
