+++ base64/lib/codec.h
